% Auto-generated: do not edit by hand
\name{dashDeckgl}

\alias{dashDeckgl}

\title{DashDeckgl component}

\description{
ExampleComponent is an example component. It takes a property, `label`, and displays it. It renders an input with the property `value` which is editable by the user.
}

\usage{
dashDeckgl(id=NULL, configuration=NULL, customLibraries=NULL,
description=NULL, events=NULL, height=NULL, overlay=NULL,
spec=NULL, tooltip=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{configuration}{Named list. }

\item{customLibraries}{Unnamed list. }

\item{description}{Named list. HTML of description elements}

\item{events}{Unnamed list. List of events to listen to}

\item{height}{Numeric. }

\item{overlay}{Character. JSON spec of the overlay deck.gl instance
(optional)}

\item{spec}{Character. JSON spec of the primary deck.gl instance}

\item{tooltip}{Named list. }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

